게시판 작업중
게시판에 제목 띄우기까지 완료.
bbs <-> bbs_content 외래키로 묶기.
글 목록 부분 마무리(클릭했을때 글 뜨도록, 조회수, 댓글수 기능 추가)
게시판쪽 DB 약간 수정
DB부분 도움 필요(외래키)

+++ loco/src/main/java/com/yuhan/loco/bbs/PostDB.java
@@ -21,7 +21,7 @@ import jakarta.persistence.Table;
 public class PostDB {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String category;
 	private String title;
     private String writer; // 작성자
@@ -29,10 +29,11 @@ public class PostDB {
     private String comment;
     
 	
-	public int getId() {
+
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getCategory() {
